Visualizaiton of Penguin in the Palmer Archipelago
The purpose of this data analysis report is to visualize and explore the relationships between various variables of penguin samples from the Palmer Archipelago. The analysis aims to investigate the relationships between factors such as species, gender, island location, and the physical attributes of penguins, including body weight, bill length and depth, and flipper length.